\name{ggpca-package}
\alias{ggpca-package}
\alias{ggpca}
\docType{package}
\title{
ggplot for pca
}
\description{
Creates ggplot biplot methods for pca objects
}
\details{
\tabular{ll}{
Package: \tab ggpca\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-02-01\cr
License: \tab GPL-3\cr
}
}
\author{Etienne Low-Décarie
}
\references{

}
\keyword{ package }
\seealso{
}
\examples{
}
