% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPCA.R
\name{ggPCA}
\alias{ggPCA}
\title{PCA plot based on ggplot2 using the stats::prcomp function}
\usage{
ggPCA(data, pc = 12, level = 0.68)
}
\arguments{
\item{data}{A data frame containing the first collumn as the grouping information}

\item{pc}{Which PC to present, one of 12, 13, 23, the default value is 12}

\item{level}{default value is 0.68}
}
\value{

}
\description{
PCA plot based on ggplot2 using the stats::prcomp function
}
\examples{
data <- iris
data <- data[, c(5, 1:4)]
colnames(data)[1] <- "group"
ggPCA(data)
}
\author{
ZhonghuiGai
}
